pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-transport-operators"
version = "0.1.0"
description = "Hybrid neural-operator surrogates for 2D advection–diffusion transport modeling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "torch>=2.1",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]
